// src/types/ItemCardProps.ts
export type ItemCardProps = {
  imageUrl: string;
  titleImage: string;
  width: number;
  height: number;
  title: string;
  description: string;
  buttonTitle?: string;
  containerStyles?: string;
  descriptionStyles?: string;
  headingStyles?: string;
  showButton?: boolean;
  price?: string;
  showPrice?: boolean;
  priceStyles?: string;
  buttonStyles?: string;
};